Added capybara312 RGB! plugin.lua
Of course, added Capybara 312 an RGB lighting. Very proud of it, lol. It's really just a few lines changed and added, but seeing it glow is always nice to see. Very colorful.

Made some major changes to the files, removed most of them. I did some testing and nothing seems to have relied on them, but if anything is broken (aside from Frames Setup) lemme know.